requirements.txt
futu-api>=7.1.3300
python-dotenv>=1.0.0
pandas>=1.3.0 